Tag EC2 Volume's with their Drive Letters
This Script add Drive Name tag to the Volumes attached to an Instance.
For Windows it will get all the Drive Names like C, D, E etc. and add this Drive Letter to the Volume associated to it
Same is the case for the linux, it will get the directory name on which the volume is mounted and tag that volume with that directory name.